+++ stories/layoutManager.stories.tsx
@@ -2,11 +2,10 @@ import { Meta } from '@storybook/react';
 import React from 'react';
 
 import { Toolbar } from '../src';
-import LayoutManager, { Layout } from '../src/layout/LayoutManager';
+import LayoutManager, { Layout } from '../src/components/LayoutManager';
 
 export default {
   title: 'Layout/LayoutManager',
-  component: fromLayoutObject,
 } as Meta;
 
 const components = {